+++ src/main/java/kitchenpos/menu/dto/MenuRequest.java
@@ -7,21 +7,21 @@ import kitchenpos.menu.domain.MenuGroup;
 
 public class MenuRequest {
     private String name;
-    private int price;
+    private Long price;
     private Long menuGroupId;
     private List<MenuProductRequest> menuProducts;
     
     private MenuRequest() {
     }
 
-    private MenuRequest(String name, int price, Long menuGroupId, List<MenuProductRequest> menuProducts) {
+    private MenuRequest(String name, Long price, Long menuGroupId, List<MenuProductRequest> menuProducts) {
         this.name = name;
         this.price = price;
         this.menuGroupId = menuGroupId;
         this.menuProducts = menuProducts;
     }
     
-    public static MenuRequest of(String name, int price, Long menuGroupId, List<MenuProductRequest> menuProducts) {
+    public static MenuRequest of(String name, Long price, Long menuGroupId, List<MenuProductRequest> menuProducts) {
         return new MenuRequest(name, price, menuGroupId, menuProducts);
     }
     
@@ -33,7 +33,7 @@ public class MenuRequest {
         return name;
     }
 
-    public int getPrice() {
+    public Long getPrice() {
         return price;
     }
